Style produto buscado
